--- a/goblint.opam
+++ b/goblint.opam
@@ -48,6 +48,7 @@ depends: [
   "conf-ruby" {with-test}
   "benchmark" {with-test}
   "conf-gcc"
+  "iter"
 ]
 depopts: ["apron" "z3"]
 conflicts: [