--- a/goblint.opam.locked
+++ b/goblint.opam.locked
@@ -101,6 +101,7 @@ depends: [
   "yaml" {= "3.1.0"}
   "yojson" {= "2.0.1"}
   "zarith" {= "1.12"}
+  "iter" {>= "1.0"}
 ]
 build: [
   ["dune" "subst"] {dev}