--- a/make.sh
+++ b/make.sh
@@ -16,7 +16,10 @@ opam_setup_flambda() {
   opam init -y -a --bare $SANDBOXING # sandboxing is disabled in travis and docker
   opam update
   # Note: the `--update-invariant` option is needed for replacement of the previous ocaml compiler switch invariant of Goblint
-  opam switch -y create . --deps-only ocaml-variants.4.14.0+options ocaml-options-only-flambda --locked --update-invariant
+  opam switch -y create . --deps-only ocaml-variants.4.14.0+options ocaml-option-flambda --locked --update-invariant
+  # Run eval here, so that we don't forget to run it manually
+  cwd = $(pwd)
+  eval $(opam env --switch=$cwd)
 }
 
 rule() {