### PR introducing the Iter package

**This PR should be merged after the flambda PR #928 and the int-domain-optimization PR #943** 

We introduce the `Iter` package [https://github.com/c-cube/iter/](url) to replace certain `List` methods within `src/analyses/base.ml` and `src/solvers/generic.ml`

We performed a sanity check using `make test` that passed without errors. Additionally, we did some benchmarking using the following file/flags:
`./goblint ../bench/coreutils/ls_comb.c --set pre.cppflags[+] "--std=gnu89" --disable ana.base.context.non-ptr --disable ana.int.def_exc --disable sem.unknown_function.spawn --set ana.thread.domain plain --enable exp.earlyglobs --set ana.base.privatization none --set pre.cppflags[+] -DGOBLINT_NO_BSEARCH --set pre.cppflags[+] -DGOBLINT_NO_ASSERT --set result json-messages --set ana.activated "['base', 'mallocWrapper']" --set ana.ctx_insens[+] base --set ana.ctx_insens[+] mallocWrapper --enable dbg.timing.enabled --enable ana.int.interval --enable ana.int.interval_threshold_widening`

Master branch goblint:
-   1229.096s -> base line

Flambda goblint:
-   1088.269s -> speed up: 11.46%
-   1109.058s -> speed up:  9.77%
-   1146.942s -> speed up:  6.68%

Iter + flambda goblint:  
-   881.201s -> speed up: 28.30%
-   913.384s -> speed up: 25.68% 
-   998.245s -> speed up: 18.78%

The measurements are done on a local machine. All speed ups are compared to the master branch goblint version.

Edit:
We also created a seperate branch called `seq-optimization` ([https://github.com/adelavais/analyzer/tree/seq-optimization](url)) that implements the same changes using OCaml's built-in `Seq` package. We'll perform benchmarks similar to the one mentioned above and compare the results with the one from this PR. Afterwards we can decide whether the `Iter` or the `Seq` package is better suited for goblint.