@montrie 2023-01-12
After some more benchmarking on my local machine, I get the following runtimes for a similar Seq based implementation as seen in #965. The measurements were performed as mentioned above:

Seq + flambda goblint (compared to the baseline measurement of 1229.096s):

- 906.564s -> speedup: 26.24%
- 977.171s -> speedup: 20.50%
- 1073.392s -> speedup: 12.67%
- 1094.149s -> speedup: 10.98%


### Runtimes on the LRZ cluster
We did an additional measurement of the Iter and Seq implementation on the **LRZ cluster**. The measured runtimes are as follows:

- Iter + flambda: 1177,313s
- Seq + flambda: 1173,092s

The runtimes of the Iter and the Seq packages on the cluster are pretty much the same. As per [http://ocamlverse.net/content/iterators.html](url) however, we would expect the combination of `Iter` and flambda's optimization flag `-O3` to achieve better performances than than `Seq` + `-O3`