@montrie 2023-01-13
Based on the above measurements, there is _no obvious choice_ for which package to use at the moment. Nevertheless, there are two points to consider:

1. **Is it worthwhile to require the installation of the Iter package given the current measurements?** According to [http://ocamlverse.net/content/iterators.html](url), `Iter` _should_ give us better performances in combination with flambda's `-O3` optimization.
2. **Will goblint implement more complex/longer chains of List manipulation operations?** Since Iter omits data allocation for intermediate List creation in a chain of `Iter` function calls, we can assume that Iter would show better performances in certain cases.