@sim642 2023-01-16
If performance-wise they should be more or less the same, then at least I would slightly prefer `Seq` as a more standard thing. We already use `BatEnum` to avoid intermediate lists in many places and `Seq` is essentially the same. It might even be good to slowly move to using `Seq` at those places as well.

Iter is conceptually more efficient, but if the `Seq` overhead is irrelevant in our usage, it might not be worth it to introduce yet another style of iteration in the Goblint codebase.