@montrie 2023-01-16
> Iter is conceptually more efficient, but if the `Seq` overhead is irrelevant in our usage, it might not be worth it to introduce yet another style of iteration in the Goblint codebase.

That's what I'm thinking as well, it doesn't seem worth it _at the moment_ to introduce the `Iter` package to goblint.