@montrie 2023-01-18
> This PR uses a different version of the configuration optimization compared to #965. It would make sense to decide what changes to the configuration handling one would like to have, and keep them in a separate PR.

Yes, this PR is based on a version of #965 where we hadn't yet decided how to do the configuration handling. We'll have that sorted out and adjust this PR accordingly.