@montrie 2023-01-28
I'll close this PR in favor of a cleaner version in #974.